Refactor: изменил двойные ковычки на одинарные и перенес из мидлвара роуты обновления и удаления пользователя для политик

+++ routes/web.php
@@ -1,6 +1,7 @@
 <?php
 
 
+use App\Http\Controllers\Admin\DestroyUserController;
 use App\Http\Controllers\Admin\UpdatePasswordUserController;
 use App\Http\Controllers\Admin\UpdateUserController;
 use App\Http\Controllers\Auth\DashboardController;
@@ -26,8 +27,6 @@ Route::middleware(['auth'])->group(function () {
     Route::get('/dashboard', DashboardController::class)->name('dashboard');
     Route::get('/profile', ProfileController::class)->name('profile');
     Route::get('/profile/edit/{user}', EditProfileController::class)->name('profile.edit');
-//    Route::patch('/profile/update/{user}', EditProfileController::class)->name('profile.update');
-//    Route::patch('/profile/update_password/{user}', EditProfileController::class)->name('profile.update.password');
 });
 
 Route::middleware(AdminPanelMiddleware::class)->namespace('App\Http\Controllers\admin')->prefix('admin')->group(function () {
@@ -36,8 +35,9 @@ Route::middleware(AdminPanelMiddleware::class)->namespace('App\Http\Controllers\
     Route::get('/create', CreateUserController::class)->name('admin.create');
     Route::post('/store', StoreUserController::class)->name('admin.store');
     Route::get('/edit/{user}', EditUserController::class)->name('admin.edit');
-    Route::delete('/destroy/{user}', DestroyUserController::class)->name('admin.destroy');
 });
 Route::patch('/admin/update/{user}', UpdateUserController::class)->name('admin.update');
 Route::patch('/admin/update_password/{user}', UpdatePasswordUserController::class)->name('admin.update_password');
+Route::delete('/admin/destroy/{user}', DestroyUserController::class)->name('admin.destroy');
+
